[app:Layout] Move the `Layout` component from `pages/Layout` to `app/Layout`
[app:dataRoutes] Update references to `Layout` component
